+++ manage.py
@@ -1,4 +1,4 @@
-from app import app, db
+from app import app,db 
 
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
@@ -9,4 +9,4 @@ manager = Manager(app)
 manager.add_command('db', MigrateCommand)
 
 if __name__ == '__main__':
-	manager.run()
+    manager.run()
