+++ project/templates/errors.html
@@ -1,0 +1,8 @@
+{% extends 'base.html' %}
+
+{% block content %}
+	<p>{{error}}</p>	
+
+	<div style="width:100%;height:0;padding-bottom:56%;position:relative;"><iframe src="https://giphy.com/embed/nVTa8D8zJUc2A" width="100%" height="100%" style="position:absolute" frameBorder="0" class="giphy-embed" allowFullScreen></iframe></div><p><a href="https://giphy.com/gifs/nVTa8D8zJUc2A">via GIPHY</a></p>
+
+{% endblock %}
