--- templates/base.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title>User Messages</title>
-</head>
-<body>
-	{% block content %}
-	{% endblock %}
-</body>
-</html>
