--- templates/messages/index.html
@@ -1,19 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-	<a href="{{url_for('messages_new', user_id=user.id)}}">Add a message for this user</a>
-	<h1>See all messages for {{user.first_name}}</h1>
-
-	{% for message in user.messages %}
-	<p>
-		{{message.content}}
-		<br>
-		<a href="{{url_for('messages_edit', user_id=user.id, id=message.id)}}">Edit a message</a>
-		<form method="POST" action="{{url_for('messages_show', user_id=user.id, id=message.id)}}?_method=DELETE">
-			{{ delete_form.hidden_tag()}}
-			<button>X</button>
-		</form>
-	</p>
-	{% endfor %}
-	
-{% endblock %}
