--- templates/messages/new.html
@@ -1,23 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-	<h1>Add a new message!</h1>
-	<form method="POST" action="{{url_for('messages_index', user_id=user.id)}}">
-		{{form.hidden_tag()}}
-
-		{% for field in form if field.widget.input_type != 'hidden' %}
-			<p>
-				{{field.label}}
-				{{field}}
-				<span>
-					{% if field.errors %}
-						{% for error in field.errors %}
-							{{ error }}
-						{% endfor %}
-					{% endif %}
-				</span>
-			</p>
-		{% endfor %}
-		<input type="submit" value='Add a new message!'>
-	</form>
-{% endblock %}
