--- templates/users/edit.html
@@ -1,24 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-	<h1>Edit a user</h1>
-
-	<form method="POST" action="{{url_for('show', id=user.id)}}?_method=PATCH">
-		{{form.hidden_tag()}}
-
-		{% for field in form if field.widget.input_type != 'hidden' %}
-			<p>
-				{{field.label}}
-				{{field}}
-				<span>
-					{% if field.errors %}
-						{% for error in field.errors %}
-							{{ error }}
-						{% endfor %}
-					{% endif %}
-				</span>
-			</p>
-		{% endfor %}
-		<button>Edit this user</button>
-	</form>
-{% endblock %}
